Refactored NavBar into separate component
